ADC & UART Configuration

+++ MCAL/UART/UART.h
@@ -135,7 +135,7 @@ enum Polling_Mechanism {
 void MCAL_UART_Init				(UART_TYPE_DEF* uartx, UART_Config_t* UARTCfg);
 void MCAL_UART_DeInit			(UART_TYPE_DEF* uartx);
 void MCAL_UART_GPIO_SetPins		(UART_TYPE_DEF* uartx);
-void MCAL_UART_SendData			(UART_TYPE_DEF* uartx, uint16_t* pTxBuffer, uint8_t len, enum Polling_Mechanism pollingEn);
+void MCAL_UART_SendData			(UART_TYPE_DEF* uartx, uint8_t* pTxBuffer, uint8_t len, enum Polling_Mechanism pollingEn);
 void MCAL_UART_ReceiveData         (UART_TYPE_DEF* uartx, uint16_t* pRxBuffer, enum Polling_Mechanism pollingEn);
 void MCAL_UART_Wait_TC				(UART_TYPE_DEF* uartx);
